add: list handler追加準備
