+++ share/src/model/Assemble.rs
@@ -62,3 +62,16 @@ pub struct AcAsmGetRes {
 }
 
 
+#[typeshare]
+#[derive(Deserialize, ToSchema)]
+#[serde(rename_all = "camelCase")]
+pub struct AcAsmListReq {
+    pub size: u64
+}
+
+#[typeshare]
+#[derive(Serialize, ToSchema)]
+#[serde(rename_all = "camelCase")]
+pub struct AcAsmListRes {
+    pub ac_assembles: Vec<AcAssemble>
+}
